Chore: make more general
